Develop the pie chart
