+++ src/components/routes/analytics/chartTypes.ts
@@ -1,0 +1,7 @@
+export interface ChartsProp {
+  data: {
+    name: string;
+    count: number;
+  }[];
+  yAxisTicks: number[];
+}
