+++ src/components/routes/analytics/charts/barChartsComp.tsx
@@ -9,19 +9,10 @@ import {
   Tooltip,
   Cell,
 } from "recharts";
+import { ChartsProp } from "../chartTypes";
 
-const BarCharts = () => {
+const BarCharts = ({ data, yAxisTicks }: ChartsProp) => {
   const [isWindow, setIsWindow] = useState(window.innerWidth);
-
-  const data = [
-    { name: "Jan", count: 13000 },
-    { name: "Feb", count: 31000 },
-    { name: "Mar", count: 10000 },
-    { name: "Apr", count: 38000 },
-    { name: "May", count: 19000 },
-  ];
-
-  const yAxisTicks = [0, 10000, 20000, 30000, 40000, 50000, 60000];
 
   const handleResize = () => {
     setIsWindow(window.innerWidth);
