+++ src/components/routes/analytics/index.ts
@@ -1,4 +1,5 @@
 import Analytics from "./analytics";
 import BarCharts from "./charts/barChartsComp";
+import PieChartComp from "./charts/pieChart";
 
-export { Analytics, BarCharts };
+export { Analytics, BarCharts, PieChartComp };
